+++ lib/src/load_status.dart
@@ -1,87 +1,42 @@
-part of '../fetching_state.dart';
-
-enum _LoadStatusEnums { init, loading, done, error, loadingMore }
+import 'package:fetching_state/src/enums.dart';
 
 /// A mixin to use with classes that have to do with loading status.
 /// This mixin will attach `loadStatus` attribute into that class along with it are many helper methods for easily checking and setting current `loadStatus`.
 /// There are also helper methods to render widget or perform action base on current `loadStatus` (when, whenOrElse, map, mapOrElse)
 mixin LoadStatusMixin {
-  _LoadStatusEnums loadStatus = _LoadStatusEnums.init;
-  void setLoadStatusInit() => loadStatus = _LoadStatusEnums.init;
-  void setLoadStatusLoading() => loadStatus = _LoadStatusEnums.loading;
-  void setLoadStatusLoadMore() => loadStatus = _LoadStatusEnums.loadingMore;
-  void setLoadStatusDone() => loadStatus = _LoadStatusEnums.done;
-  void setLoadStatusError() => loadStatus = _LoadStatusEnums.error;
+  LoadState loadStatus = LoadState.init;
+  void setLoadStatusInit() => loadStatus = LoadState.init;
+  void setLoadStatusLoading() => loadStatus = LoadState.loading;
+  void setLoadStatusLoadMore() => loadStatus = LoadState.loadingMore;
+  void setLoadStatusDone() => loadStatus = LoadState.done;
+  void setLoadStatusError() => loadStatus = LoadState.error;
 
-  bool get isInit => loadStatus == _LoadStatusEnums.init;
-  bool get isLoading => loadStatus == _LoadStatusEnums.loading;
-  bool get isLoadMore => loadStatus == _LoadStatusEnums.loadingMore;
-  bool get isDone => loadStatus == _LoadStatusEnums.done;
-  bool get isError => loadStatus == _LoadStatusEnums.error;
+  bool get isInit => loadStatus == LoadState.init;
+  bool get isLoading => loadStatus == LoadState.loading;
+  bool get isLoadMore => loadStatus == LoadState.loadingMore;
+  bool get isDone => loadStatus == LoadState.done;
+  bool get isError => loadStatus == LoadState.error;
 
-  Widget whenOrElse({
-    Widget Function()? onLoading,
-    Widget Function(bool isLoadingMore)? onDone,
-    Widget Function()? onError,
-    Widget Function()? onInit,
-    required Widget Function() orElse,
-  }) {
-    switch (loadStatus) {
-      case _LoadStatusEnums.init:
-        return onInit == null ? orElse() : onInit();
-      case _LoadStatusEnums.loading:
-        return onLoading == null ? orElse() : onLoading();
-      case _LoadStatusEnums.loadingMore:
-      case _LoadStatusEnums.done:
-        return onDone == null
-            ? orElse()
-            : onDone(loadStatus == _LoadStatusEnums.loadingMore);
-      case _LoadStatusEnums.error:
-        return onError == null ? orElse() : onError();
-      default:
-        return orElse();
-    }
-  }
-
-  Widget when({
-    required Widget Function() onLoading,
-    required Widget Function(bool isLoadingMore) onDone,
-    required Widget Function() onError,
-    required Widget Function() onInit,
-  }) {
-    switch (loadStatus) {
-      case _LoadStatusEnums.init:
-        return onInit();
-      case _LoadStatusEnums.loading:
-        return onLoading();
-      case _LoadStatusEnums.loadingMore:
-      case _LoadStatusEnums.done:
-        return onDone(loadStatus == _LoadStatusEnums.loadingMore);
-      case _LoadStatusEnums.error:
-        return onError();
-    }
-  }
-
-  R map<R>({
+  R when<R>({
     required R Function() onLoading,
     required R Function(bool isLoadingMore) onDone,
     required R Function() onError,
     required R Function() onInit,
   }) {
     switch (loadStatus) {
-      case _LoadStatusEnums.init:
+      case LoadState.init:
         return onInit();
-      case _LoadStatusEnums.loading:
+      case LoadState.loading:
         return onLoading();
-      case _LoadStatusEnums.loadingMore:
-      case _LoadStatusEnums.done:
-        return onDone(loadStatus == _LoadStatusEnums.loadingMore);
-      case _LoadStatusEnums.error:
+      case LoadState.loadingMore:
+      case LoadState.done:
+        return onDone(loadStatus == LoadState.loadingMore);
+      case LoadState.error:
         return onError();
     }
   }
 
-  R mapOrElse<R>({
+  R whenOrElse<R>({
     R Function()? onLoading,
     R Function(bool isLoadingMore)? onDone,
     R Function()? onError,
@@ -89,16 +44,16 @@ mixin LoadStatusMixin {
     required R Function() orElse,
   }) {
     switch (loadStatus) {
-      case _LoadStatusEnums.init:
+      case LoadState.init:
         return onInit == null ? orElse() : onInit();
-      case _LoadStatusEnums.loading:
+      case LoadState.loading:
         return onLoading == null ? orElse() : onLoading();
-      case _LoadStatusEnums.loadingMore:
-      case _LoadStatusEnums.done:
+      case LoadState.loadingMore:
+      case LoadState.done:
         return onDone == null
             ? orElse()
-            : onDone(loadStatus == _LoadStatusEnums.loadingMore);
-      case _LoadStatusEnums.error:
+            : onDone(loadStatus == LoadState.loadingMore);
+      case LoadState.error:
         return onError == null ? orElse() : onError();
       default:
         return orElse();