+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 3.0.0
+
+- remove `map` methods
+- remove `flutter` dependency
+- add [LoadingState] which is similar to [FetchingState] but without `data`. Use when you don't need to attach a data object in to the [FetchingState]
+
+## 2.0.3
+
+- small fixes
+
+## 2.0.2
+
+- in [FetchingState] make `T? data` optional.
+
 ## 2.0.1
 
 - update example and documents